+++ RELEASE_NOTES.md
@@ -1,7 +1,13 @@
 ## v0.3.0 (2022-07-22)
 
 - Completer
+  - Constructor now has a required `parent` arg, and the resultant
+    widget is appended as a child of that element
+  - `getValue` renamed `getData` for consistency with future widgets
   - `C-r` skips matches when they are the same as the previous match
+  - History loading message now suppressed unless there is history to
+    load
+  - Widgets now have a CSS classname set on them
 
 
 ## v0.2.0 (2022-07-21)